refactor(components): split admin/page.tsx and inbox/page.tsx into sub-components (P2-6)
